backend/requirements.txt
# Jobsy AI Backend Dependencies
# =============================================================================

# Core FastAPI Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Vector Database
chromadb>=0.4.0
sentence-transformers>=2.2.0

# HTTP Client
requests>=2.31.0
httpx>=0.25.0

# Environment & Configuration
python-dotenv>=1.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Web Scraping & Text Processing
beautifulsoup4>=4.12.0
langdetect>=1.0.9

# AI & ML (Optional - for advanced features)
# transformers>=4.35.0
# torch>=2.1.0
# google-generativeai>=0.3.0

# Database & ORM
# sqlalchemy>=2.0.0
# alembic>=1.12.0

# Authentication & Security
# python-jose[cryptography]>=3.3.0
# passlib[bcrypt]>=1.7.4

# Utilities
python-multipart>=0.0.6
email-validator>=2.1.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.1.0

# =============================================================================
# NOTLAR:
# 1. AI paketleri (transformers, torch) opsiyonel - sadece gerekirse kurun
# 2. Geliştirme paketleri sadece dev environment'da kurun
# 3. Versiyonlar minimum uyumluluk için belirtilmiştir
# ============================================================================= 